FIX #8 tracks with XM don't work
Tweets with an explicit mention (XM) and a track only (e.g., $1733) were
not answered. Now they are.

There is now a little more commentation in react.find_tokens, where the
bug lay:

If there is, in mode 'all' (which is active for XM tweets), only one
result, and that result is not the magic hash tag, then return it.
If it is the magic hashtag, don't return but look for non-sigil'd
results.

CHANGELOG Streckennummern in Tweets mit expliziter Erwähnung werden gefunden cc @ml_navigat