FIX #62 Testwrite to Database
In order to find mistakes in the code parts that insert data into the
database, we need to actually call that code. In readonly mode (which is
on during testing) all changes will be discared at the end anyway.

pylint score: 10.00
test score: Level 51:test_ds100:TESTS FAIL TO RUN