FIX store version status directly
If there is a crash after posting the "new version" status, the bot
until didn't remember sending it and thus sent it again in the next run,
mostly failing *again*. This, of course, is unwanted behaviour.

Now, we store and commit the version toot info directly.

pylint score: 10.00
test score: A 29/0 R 100.0%
pytest score: ................................. [100%]